refactor(#57): renamed gen->aigen
