+++ internal/app/app_test.go
@@ -7,7 +7,7 @@ import (
 	"testing"
 
 	"github.com/baudii/ada-ai/internal/app"
-	"github.com/baudii/ada-ai/internal/core/gen"
+	"github.com/baudii/ada-ai/internal/core/aigen"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -44,7 +44,7 @@ func (m *mockGen) BuildPrompt(p string, args ...any) (string, error) {
 	return m.promptStr, m.promptErr
 }
 
-func (m *mockGen) GenerateWithSys(ctx context.Context, sys, user string, opts ...gen.Option) (string, error) {
+func (m *mockGen) GenerateWithSys(ctx context.Context, sys, user string, opts ...aigen.Option) (string, error) {
 	return m.resp, m.genErr
 }
